deployment/requirements.txt
# Core FastAPI dependencies
fastapi==0.95.2
starlette==0.27.0
pydantic==1.10.12

# AWS dependencies
boto3==1.35.76
mangum==0.19.0

# Authentication dependencies
python-jose==3.3.0
ecdsa==0.19.0
rsa==4.9
pyasn1==0.6.0
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0

# HTTP client dependencies (for requests)
urllib3==1.26.18
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4

# Type hints
typing_extensions==4.8.0
